Update profiles.yml
There are a number of Data Commons-specific parameters and secrets defined in the `credentials.env` file.  The example file now shows how these parameters can be accessed using the env_var functionality of YAML files.  This has the benefit that when a user is managing a number of dbt projects over a period of time, they won't have to update things like the `jwt_token` several times every few days (which gets old quickly).  It has the nice side effect that the `profiles.yml` file no longer contains any secrets.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>